Working in game again
hooray

+++ dScripts/ai/ACT/FootRace/BaseFootRaceManager.cpp
@@ -42,7 +42,7 @@ void BaseFootRaceManager::OnFireEventServerSide(Entity* self, Entity* sender, st
 					}
 
 					StopActivity(self, player->GetObjectID(), 0, param1);
-					SaveScore(self, player->GetObjectID(), param1, param2, param3);
+					SaveScore(self, player->GetObjectID(), static_cast<float>(param1), static_cast<float>(param2), static_cast<float>(param3));
 				}
 			}
 		}
